remove IdeaDownloadView and register export in dashboard
