+++ euth/ideas/urls.py
@@ -5,8 +5,6 @@ from . import views
 urlpatterns = [
     url(r'create/module/(?P<slug>[-\w_]+)/$',
         views.IdeaCreateView.as_view(), name='idea-create'),
-    url(r'download/module/(?P<slug>[-\w_]+)/$',
-        views.IdeaDownloadView.as_view(), name='idea-download'),
     url(r'^(?P<slug>[-\w_]+)/edit/$',
         views.IdeaUpdateView.as_view(), name='idea-update'),
     url(r'^(?P<slug>[-\w_]+)/delete/$',
